ml_service/requirements.txt
kafka-python==2.0.2
psycopg2-binary==2.9.9
python-dotenv==1.0.0
scikit-learn==1.3.0
pandas==2.0.3
numpy==1.24.3
joblib==1.3.2
schedule==1.2.0 